Add files via upload
The updated version of the purification for the library build. The script includes better time and magnet position for the bead attraction (here about 1/4 up the tube, just below 40 µL in volume), and specific ethanol removal to leave no residual ethanol after the last removal step and drying. The total drying time is estimated to be 4 mins and 55 seconds, but can vary based on the run.